examples/python/requirements.txt
tempus-sdk>=0.1.0
python-dotenv>=1.0.0
